test(history-index): split pinned generation recovery tests
