+++ crates/ctx-history-index/src/tests/mod.rs
@@ -349,6 +349,7 @@ fn multisegment_fixture(
     (temp, sources)
 }
 
+mod pinned_generation;
 mod query;
 mod recovery;
 mod writer;
